pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-impact"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1",
    "anthropic>=0.25",
    "click>=8.1",
    "rich>=13.0",
]

[project.scripts]
pr-impact = "pr_impact.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pr_impact*"]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pyright>=1.1",
    "pre-commit>=3.7",
    "detect-secrets>=1.5",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (ruff format handles this)
]

[tool.ruff.lint.isort]
known-first-party = ["pr_impact"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
